client/requirements.txt
pycapnp
six
future
